requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --extra=dev
#
click==8.1.7
    # via typer
dtyper==2.4.0
    # via render-engine
gitdb==4.0.11
    # via gitpython
gitpython==3.1.40
    # via render-engine
iniconfig==2.0.0
    # via pytest
jinja2==3.1.2
    # via
    #   render-engine
    #   render-engine-microblog (pyproject.toml)
markdown-it-py==3.0.0
    # via rich
markdown2==2.4.10
    # via render-engine
markupsafe==2.1.3
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.1.0
    # via render-engine
packaging==23.2
    # via pytest
pluggy==1.3.0
    # via
    #   pytest
    #   render-engine
pygments==2.16.1
    # via rich
pytest==7.4.2
    # via
    #   pytest-mock
    #   render-engine-microblog (pyproject.toml)
pytest-mock==3.12.0
    # via render-engine-microblog (pyproject.toml)
python-dateutil==2.8.2
    # via render-engine
python-frontmatter==1.0.0
    # via render-engine
python-slugify==8.0.1
    # via render-engine
pyyaml==6.0.1
    # via python-frontmatter
render-engine==2023.10.2a4
    # via render-engine-microblog (pyproject.toml)
rich==13.6.0
    # via render-engine
six==1.16.0
    # via python-dateutil
smmap==5.0.1
    # via gitdb
text-unidecode==1.3
    # via python-slugify
typer==0.9.0
    # via
    #   dtyper
    #   render-engine
typing-extensions==4.8.0
    # via typer
